pyproject.toml
[project]
name = "ai-travel-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.82.0",
    "resonate-sdk>=0.5.3",
    "unstructured>=0.16.23",
]

[project.scripts]
app = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
